reeworkeed the stucture and it s redirect
